requirements.txt
wikipedia-api
networkx
pyvis
openai
requests
groq
python-pptx
